backend/requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
PyJWT==2.10.1
pytesseract==0.3.13
requests==2.32.5
SQLAlchemy==2.0.41
stripe==12.5.1
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
